Implement improvements to payments implementation
